qcommon: create q_math.h and transform most functions as inline ones
