+++ src/engine/qcommon/q_shared.h
@@ -37,14 +37,6 @@ Maryland 20850 USA.
 
 #include "common/Defs.h"
 
-// math.h/cmath uses _USE_MATH_DEFINES to decide if to define M_PI etc or not.
-// So define _USE_MATH_DEFINES early before including math.h/cmath
-// and before including any other header in case they bring in math.h/cmath indirectly.
-#ifndef _USE_MATH_DEFINES
-#define _USE_MATH_DEFINES
-#endif
-
-
 // q_shared.h -- included first by ALL program modules.
 // A user mod should never modify this file
 
@@ -63,19 +55,14 @@ Maryland 20850 USA.
 #define CLIENT_WINDOW_TITLE     PRODUCT_NAME
 #define CLIENT_WINDOW_MIN_TITLE PRODUCT_NAME_LOWER
 
-
-#define Q_UNUSED(x) (void)(x)
-
 template<class T>
 void ignore_result(T) {}
 
 // C standard library headers
 #include <errno.h>
-#include <float.h>
 #include <iso646.h>
 #include <limits.h>
 #include <locale.h>
-#include <math.h>
 #include <setjmp.h>
 #include <signal.h>
 #include <stdarg.h>
@@ -137,29 +124,11 @@ void ignore_result(T) {}
 #define Q_snprintf  snprintf
 #endif
 
-// msvc does not have roundf
-#ifdef _MSC_VER
-#define roundf( f ) ( floor( (f) + 0.5 ) )
-#endif
-
-//=============================================================
-
-	using byte = uint8_t;
-	using uint = unsigned int;
-	enum class qtrinary {qno, qyes, qmaybe};
-
-	union floatint_t
-	{
-		float f;
-		int i;
-		uint ui;
-	};
-
-//=============================================================
-
 #include "common/Platform.h"
 #include "common/Compiler.h"
 #include "common/Endian.h"
+
+enum class qtrinary {qno, qyes, qmaybe};
 
 using qhandle_t = int;
 using sfxHandle_t = int;
@@ -227,101 +196,7 @@ using clipHandle_t = int;
 MALLOC_LIKE void *Com_Allocate_Aligned( size_t alignment, size_t size );
 void  Com_Free_Aligned( void *ptr );
 
-	/*
-	==============================================================
-
-	MATHLIB
-
-	==============================================================
-	*/
-
-	using vec_t = float;
-	using vec2_t = vec_t[2];
-
-	using vec3_t = vec_t[3];
-	using vec4_t = vec_t[4];
-
-	using axis_t = vec3_t[3];
-	using matrix3x3_t = vec_t[3 * 3];
-	using matrix_t = vec_t[4 * 4];
-	using quat_t = vec_t[4];
-
-	// A transform_t represents a product of basic
-	// transformations, which are a rotation about an arbitrary
-	// axis, a uniform scale or a translation. Any a product can
-	// alway be brought into the form rotate, then scale, then
-	// translate. So the whole transform_t can be stored in 8
-	// floats (quat: 4, scale: 1, translation: 3), which is very
-	// convenient for SSE and GLSL, which operate on 4-dimensional
-	// float vectors.
-#if idx86_sse
-    // Here we have a union of scalar struct and sse struct, transform_u and the
-    // scalar struct must match transform_t so we have to use anonymous structs.
-    // We disable compiler warnings when using -Wpedantic for this specific case.
-#ifdef __GNUC__
-#pragma GCC diagnostic push
-#pragma GCC diagnostic ignored "-Wpedantic"
-#endif
-	ALIGNED(16, union transform_t {
-		struct {
-			quat_t rot;
-			vec3_t trans;
-			vec_t  scale;
-		};
-		struct {
-			__m128 sseRot;
-			__m128 sseTransScale;
-		};
-	});
-#ifdef __GNUC__
-#pragma GCC diagnostic pop
-#endif
-#else
-	ALIGNED(16, struct transform_t {
-		quat_t rot;
-		vec3_t trans;
-		vec_t  scale;
-	});
-#endif
-
-	using fixed4_t = int;
-	using fixed8_t = int;
-	using fixed16_t = int;
-
-#ifndef M_PI
-#define M_PI 3.14159265358979323846f // matches value in gcc v2 math.h
-#endif
-
-#ifndef M_SQRT2
-#define M_SQRT2 1.414213562f
-#endif
-
-#ifndef M_ROOT3
-#define M_ROOT3 1.732050808f
-#endif
-
-#ifndef LINE_DISTANCE_EPSILON
-#define LINE_DISTANCE_EPSILON 1e-05f
-#endif
-
 #define ARRAY_LEN(x) ( sizeof( x ) / sizeof( *( x ) ) )
-
-// angle indexes
-#define PITCH 0 // up / down
-#define YAW   1 // left / right
-#define ROLL  2 // fall over
-
-// plane sides
-	enum class planeSide_t : int
-	{
-	  SIDE_FRONT = 0,
-	  SIDE_BACK = 1,
-	  SIDE_ON = 2,
-	  SIDE_CROSS = 3
-	};
-
-#define NUMVERTEXNORMALS 162
-	extern vec3_t bytedirs[ NUMVERTEXNORMALS ];
 
 // all drawing is done to a 640*480 virtual screen size
 // and will be automatically scaled to the real resolution
@@ -342,996 +217,7 @@ void  Com_Free_Aligned( void *ptr );
 
 #include "logging.h"
 
-#define DEG2RAD( a )                  ( ( ( a ) * M_PI ) / 180.0f )
-#define RAD2DEG( a )                  ( ( ( a ) * 180.0f ) / M_PI )
-
-struct cplane_t;
-
-extern vec3_t   vec3_origin;
-extern vec3_t   axisDefault[ 3 ];
-extern matrix_t matrixIdentity;
-extern quat_t   quatIdentity;
-
-#define nanmask ( 255 << 23 )
-
-#define IS_NAN( x ) ( ( ( *(int *)&( x ) ) & nanmask ) == nanmask )
-
-#define Q_ftol(x) ((long)(x))
-
-	inline unsigned int Q_floatBitsToUint( float number )
-	{
-		floatint_t t;
-
-		t.f = number;
-		return t.ui;
-	}
-
-	inline float Q_uintBitsToFloat( unsigned int number )
-	{
-		floatint_t t;
-
-		t.ui = number;
-		return t.f;
-	}
-
-	inline float Q_rsqrt( float number )
-	{
-		float x = 0.5f * number;
-		float y;
-
-		Q_UNUSED(x);
-
-		// compute approximate inverse square root
-#if defined( idx86_sse )
-		_mm_store_ss( &y, _mm_rsqrt_ss( _mm_load_ss( &number ) ) );
-#elif idppc
-
-#ifdef __GNUC__
-		asm( "frsqrte %0, %1" : "=f"( y ) : "f"( number ) );
-#else
-		y = __frsqrte( number );
-#endif
-#else
-		y = Q_uintBitsToFloat( 0x5f3759df - (Q_floatBitsToUint( number ) >> 1) );
-		y *= ( 1.5f - ( x * y * y ) ); // initial iteration
-#endif
-		y *= ( 1.5f - ( x * y * y ) ); // second iteration for higher precision
-		return y;
-	}
-
-inline float Q_fabs( float x )
-{
-	return fabsf( x );
-}
-
-byte         ClampByte( int i );
-signed char  ClampChar( int i );
-
-// this isn't a real cheap function to call!
-int          DirToByte( vec3_t dir );
-void         ByteToDir( int b, vec3_t dir );
-
-#define DotProduct( x,y )            ( ( x )[ 0 ] * ( y )[ 0 ] + ( x )[ 1 ] * ( y )[ 1 ] + ( x )[ 2 ] * ( y )[ 2 ] )
-#define VectorSubtract( a,b,c )      ( ( c )[ 0 ] = ( a )[ 0 ] - ( b )[ 0 ],( c )[ 1 ] = ( a )[ 1 ] - ( b )[ 1 ],( c )[ 2 ] = ( a )[ 2 ] - ( b )[ 2 ] )
-#define VectorAdd( a,b,c )           ( ( c )[ 0 ] = ( a )[ 0 ] + ( b )[ 0 ],( c )[ 1 ] = ( a )[ 1 ] + ( b )[ 1 ],( c )[ 2 ] = ( a )[ 2 ] + ( b )[ 2 ] )
-#define VectorCopy( a,b )            ( ( b )[ 0 ] = ( a )[ 0 ],( b )[ 1 ] = ( a )[ 1 ],( b )[ 2 ] = ( a )[ 2 ] )
-#define VectorScale( v, s, o )       ( ( o )[ 0 ] = ( v )[ 0 ] * ( s ),( o )[ 1 ] = ( v )[ 1 ] * ( s ),( o )[ 2 ] = ( v )[ 2 ] * ( s ) )
-/** Stands for MultiplyAdd: adding a vector "b" scaled by "s" to "v" and writing it to "o" */
-#define VectorMA( v, s, b, o )       ( ( o )[ 0 ] = ( v )[ 0 ] + ( b )[ 0 ] * ( s ),( o )[ 1 ] = ( v )[ 1 ] + ( b )[ 1 ] * ( s ),( o )[ 2 ] = ( v )[ 2 ] + ( b )[ 2 ] * ( s ) )
-#define VectorLerpTrem( f, s, e, r ) (( r )[ 0 ] = ( s )[ 0 ] + ( f ) * (( e )[ 0 ] - ( s )[ 0 ] ), \
-                                      ( r )[ 1 ] = ( s )[ 1 ] + ( f ) * (( e )[ 1 ] - ( s )[ 1 ] ), \
-                                      ( r )[ 2 ] = ( s )[ 2 ] + ( f ) * (( e )[ 2 ] - ( s )[ 2 ] ))
-
-#define VectorClear( a )             ( ( a )[ 0 ] = ( a )[ 1 ] = ( a )[ 2 ] = 0 )
-#define VectorNegate( a,b )          ( ( b )[ 0 ] = -( a )[ 0 ],( b )[ 1 ] = -( a )[ 1 ],( b )[ 2 ] = -( a )[ 2 ] )
-#define VectorSet( v, x, y, z )      ( ( v )[ 0 ] = ( x ), ( v )[ 1 ] = ( y ), ( v )[ 2 ] = ( z ) )
-
-#define Vector2Set( v, x, y )        ( ( v )[ 0 ] = ( x ),( v )[ 1 ] = ( y ) )
-#define Vector2Copy( a,b )           ( ( b )[ 0 ] = ( a )[ 0 ],( b )[ 1 ] = ( a )[ 1 ] )
-#define Vector2Subtract( a,b,c )     ( ( c )[ 0 ] = ( a )[ 0 ] - ( b )[ 0 ],( c )[ 1 ] = ( a )[ 1 ] - ( b )[ 1 ] )
-
-#define Vector4Set( v, x, y, z, n )  ( ( v )[ 0 ] = ( x ),( v )[ 1 ] = ( y ),( v )[ 2 ] = ( z ),( v )[ 3 ] = ( n ) )
-#define Vector4Copy( a,b )           ( ( b )[ 0 ] = ( a )[ 0 ],( b )[ 1 ] = ( a )[ 1 ],( b )[ 2 ] = ( a )[ 2 ],( b )[ 3 ] = ( a )[ 3 ] )
-/** Stands for MultiplyAdd: adding a vector "b" scaled by "s" to "v" and writing it to "o" */
-#define Vector4MA( v, s, b, o )      ( ( o )[ 0 ] = ( v )[ 0 ] + ( b )[ 0 ] * ( s ),( o )[ 1 ] = ( v )[ 1 ] + ( b )[ 1 ] * ( s ),( o )[ 2 ] = ( v )[ 2 ] + ( b )[ 2 ] * ( s ),( o )[ 3 ] = ( v )[ 3 ] + ( b )[ 3 ] * ( s ) )
-#define Vector4Average( v, b, s, o ) ( ( o )[ 0 ] = ( ( v )[ 0 ] * ( 1 - ( s ) ) ) + ( ( b )[ 0 ] * ( s ) ),( o )[ 1 ] = ( ( v )[ 1 ] * ( 1 - ( s ) ) ) + ( ( b )[ 1 ] * ( s ) ),( o )[ 2 ] = ( ( v )[ 2 ] * ( 1 - ( s ) ) ) + ( ( b )[ 2 ] * ( s ) ),( o )[ 3 ] = ( ( v )[ 3 ] * ( 1 - ( s ) ) ) + ( ( b )[ 3 ] * ( s ) ) )
-
-#define DotProduct4(x, y)            (( x )[ 0 ] * ( y )[ 0 ] + ( x )[ 1 ] * ( y )[ 1 ] + ( x )[ 2 ] * ( y )[ 2 ] + ( x )[ 3 ] * ( y )[ 3 ] )
-
-#define SnapVector( v )              do { ( v )[ 0 ] = ( floor( ( v )[ 0 ] + 0.5f ) ); ( v )[ 1 ] = ( floor( ( v )[ 1 ] + 0.5f ) ); ( v )[ 2 ] = ( floor( ( v )[ 2 ] + 0.5f ) ); } while ( 0 )
-
-	float    RadiusFromBounds( const vec3_t mins, const vec3_t maxs );
-	void     ZeroBounds( vec3_t mins, vec3_t maxs );
-	void     ClearBounds( vec3_t mins, vec3_t maxs );
-	void     AddPointToBounds( const vec3_t v, vec3_t mins, vec3_t maxs );
-
-	void     BoundsAdd( vec3_t mins, vec3_t maxs, const vec3_t mins2, const vec3_t maxs2 );
-	bool BoundsIntersect( const vec3_t mins, const vec3_t maxs, const vec3_t mins2, const vec3_t maxs2 );
-	bool BoundsIntersectSphere( const vec3_t mins, const vec3_t maxs, const vec3_t origin, vec_t radius );
-	bool BoundsIntersectPoint( const vec3_t mins, const vec3_t maxs, const vec3_t origin );
-	float BoundsMaxExtent( const vec3_t mins, const vec3_t maxs );
-
-	inline void BoundsToCorners( const vec3_t mins, const vec3_t maxs, vec3_t corners[ 8 ] )
-	{
-		VectorSet( corners[ 0 ], mins[ 0 ], maxs[ 1 ], maxs[ 2 ] );
-		VectorSet( corners[ 1 ], maxs[ 0 ], maxs[ 1 ], maxs[ 2 ] );
-		VectorSet( corners[ 2 ], maxs[ 0 ], mins[ 1 ], maxs[ 2 ] );
-		VectorSet( corners[ 3 ], mins[ 0 ], mins[ 1 ], maxs[ 2 ] );
-		VectorSet( corners[ 4 ], mins[ 0 ], maxs[ 1 ], mins[ 2 ] );
-		VectorSet( corners[ 5 ], maxs[ 0 ], maxs[ 1 ], mins[ 2 ] );
-		VectorSet( corners[ 6 ], maxs[ 0 ], mins[ 1 ], mins[ 2 ] );
-		VectorSet( corners[ 7 ], mins[ 0 ], mins[ 1 ], mins[ 2 ] );
-	}
-
-	int VectorCompare( const vec3_t v1, const vec3_t v2 );
-
-	inline int Vector4Compare( const vec4_t v1, const vec4_t v2 )
-	{
-		if ( v1[ 0 ] != v2[ 0 ] || v1[ 1 ] != v2[ 1 ] || v1[ 2 ] != v2[ 2 ] || v1[ 3 ] != v2[ 3 ] )
-		{
-			return 0;
-		}
-
-		return 1;
-	}
-
-	inline void VectorLerp( const vec3_t from, const vec3_t to, float frac, vec3_t out )
-	{
-		out[ 0 ] = from[ 0 ] + ( ( to[ 0 ] - from[ 0 ] ) * frac );
-		out[ 1 ] = from[ 1 ] + ( ( to[ 1 ] - from[ 1 ] ) * frac );
-		out[ 2 ] = from[ 2 ] + ( ( to[ 2 ] - from[ 2 ] ) * frac );
-	}
-
-	inline int VectorCompareEpsilon( const vec3_t v1, const vec3_t v2, float epsilon )
-	{
-		vec3_t d;
-
-		VectorSubtract( v1, v2, d );
-		d[ 0 ] = fabs( d[ 0 ] );
-		d[ 1 ] = fabs( d[ 1 ] );
-		d[ 2 ] = fabs( d[ 2 ] );
-
-		if ( d[ 0 ] > epsilon || d[ 1 ] > epsilon || d[ 2 ] > epsilon )
-		{
-			return 0;
-		}
-
-		return 1;
-	}
-
-	inline void VectorMin(const vec3_t a, const vec3_t b, vec3_t out)
-	{
-		out[0] = a[0] < b[0] ? a[0] : b[0];
-		out[1] = a[1] < b[1] ? a[1] : b[1];
-		out[2] = a[2] < b[2] ? a[2] : b[2];
-	}
-
-	inline void VectorMax(const vec3_t a, const vec3_t b, vec3_t out)
-	{
-		out[0] = a[0] > b[0] ? a[0] : b[0];
-		out[1] = a[1] > b[1] ? a[1] : b[1];
-		out[2] = a[2] > b[2] ? a[2] : b[2];
-	}
-
-	vec_t VectorLength( const vec3_t v );
-	vec_t VectorLengthSquared( const vec3_t v );
-	vec_t Distance( const vec3_t p1, const vec3_t p2 );
-	vec_t DistanceSquared( const vec3_t p1, const vec3_t p2 );
-	void  CrossProduct( const vec3_t v1, const vec3_t v2, vec3_t cross );
-	vec_t VectorNormalize( vec3_t v );  // returns vector length
-	void  VectorNormalizeFast( vec3_t v );  // does NOT return vector length, uses rsqrt approximation
-	vec_t VectorNormalize2( const vec3_t v, vec3_t out );
-	void  VectorInverse( vec3_t v );
-
-	int   NearestPowerOfTwo( int val );
-
-	int   Q_rand( int *seed );
-	float Q_random( int *seed );
-	float Q_crandom( int *seed );
-
-#define random()  ( ( rand() & 0x7fff ) / ( (float)0x7fff ) )
-#define crandom() ( 2.0 * ( random() - 0.5 ) )
-
-	void vectoangles( const vec3_t value1, vec3_t angles );
-
-	void  AnglesToAxis( const vec3_t angles, vec3_t axis[ 3 ] );
-// TTimo: const vec_t ** would require explicit casts for ANSI C conformance
-// see unix/const-arg.c
-	void  AxisToAngles( /*const*/ vec3_t axis[ 3 ], vec3_t angles );
-//void AxisToAngles ( const vec3_t axis[3], vec3_t angles );
-	float VectorDistanceSquared( vec3_t v1, vec3_t v2 );
-
-	void  AxisClear( vec3_t axis[ 3 ] );
-	void  AxisCopy( vec3_t in[ 3 ], vec3_t out[ 3 ] );
-
-	void  SetPlaneSignbits( struct cplane_t *out );
-	int   BoxOnPlaneSide( const vec3_t emins, const vec3_t emaxs, const struct cplane_t *plane );
-
-	float AngleMod( float a );
-	float LerpAngle( float from, float to, float frac );
-	float AngleSubtract( float a1, float a2 );
-	void  AnglesSubtract( vec3_t v1, vec3_t v2, vec3_t v3 );
-
-	float AngleNormalize360( float angle );
-	float AngleNormalize180( float angle );
-	float AngleDelta( float angle1, float angle2 );
-	float AngleBetweenVectors( const vec3_t a, const vec3_t b );
-	void  AngleVectors( const vec3_t angles, vec3_t forward, vec3_t right, vec3_t up );
-
-	vec_t PlaneNormalize( vec4_t plane );  // returns normal length
-
-	/* greebo: This calculates the intersection point of three planes.
-	 * Returns <0,0,0> if no intersection point could be found, otherwise returns the coordinates of the intersection point
-	 * (this may also be 0,0,0) */
-	bool PlanesGetIntersectionPoint( const vec4_t plane1, const vec4_t plane2, const vec4_t plane3, vec3_t out );
-	void     PlaneIntersectRay( const vec3_t rayPos, const vec3_t rayDir, const vec4_t plane, vec3_t res );
-
-	bool PlaneFromPoints( vec4_t plane, const vec3_t a, const vec3_t b, const vec3_t c );
-	bool PlaneFromPointsOrder( vec4_t plane, const vec3_t a, const vec3_t b, const vec3_t c, bool cw );
-	void     ProjectPointOnPlane( vec3_t dst, const vec3_t point, const vec3_t normal );
-	void     RotatePointAroundVector( vec3_t dst, const vec3_t dir, const vec3_t point, float degrees );
-
-	void     RotateAroundDirection( vec3_t axis[ 3 ], float yaw );
-	void     MakeNormalVectors( const vec3_t forward, vec3_t right, vec3_t up );
-
-	float    ProjectPointOntoRectangleOutwards( vec2_t out, const vec2_t point, const vec2_t dir, const vec2_t bounds[ 2 ] );
-	void     ExponentialFade( float *value, float target, float lambda, float timedelta );
-	#define  LinearRemap(x,an,ap,bn,bp) (((x)-(an))/((ap)-(an))*((bp)-(bn))+(bn))
-
-// perpendicular vector could be replaced by this
-
-//int       PlaneTypeForNormal( vec3_t normal );
-
-	void VectorMatrixMultiply( const vec3_t p, vec3_t m[ 3 ], vec3_t out );
-
-// RB: NOTE renamed MatrixMultiply to AxisMultiply because it conflicts with most new matrix functions
-// It is important for mod developers to do this change as well or they risk a memory corruption by using
-// the other MatrixMultiply function.
-	void  AxisMultiply( float in1[ 3 ][ 3 ], float in2[ 3 ][ 3 ], float out[ 3 ][ 3 ] );
-	void  PerpendicularVector( vec3_t dst, const vec3_t src );
-
-// Ridah
-	void  GetPerpendicularViewVector( const vec3_t point, const vec3_t p1, const vec3_t p2, vec3_t up );
-	void  ProjectPointOntoVector( vec3_t point, vec3_t vStart, vec3_t vEnd, vec3_t vProj );
-	void  ProjectPointOntoVectorBounded( vec3_t point, vec3_t vStart, vec3_t vEnd, vec3_t vProj );
-	float DistanceFromLineSquared( vec3_t p, vec3_t lp1, vec3_t lp2 );
-
-// done.
-
-	vec_t DistanceBetweenLineSegmentsSquared( const vec3_t sP0, const vec3_t sP1,
-	    const vec3_t tP0, const vec3_t tP1, float *s, float *t );
-
-//=============================================
-
-// RB: XreaL matrix math functions required by the renderer
-
-	void     MatrixIdentity( matrix_t m );
-	void     MatrixClear( matrix_t m );
-	void     MatrixCopy( const matrix_t in, matrix_t out );
-	bool MatrixCompare( const matrix_t a, const matrix_t b );
-	void     MatrixTransposeIntoXMM( const matrix_t m );
-	void     MatrixTranspose( const matrix_t in, matrix_t out );
-
-// invert any m4x4 using Kramer's rule.. return true if matrix is singular, else return false
-	bool MatrixInverse( matrix_t m );
-	void     MatrixSetupXRotation( matrix_t m, vec_t degrees );
-	void     MatrixSetupYRotation( matrix_t m, vec_t degrees );
-	void     MatrixSetupZRotation( matrix_t m, vec_t degrees );
-	void     MatrixSetupTranslation( matrix_t m, vec_t x, vec_t y, vec_t z );
-	void     MatrixSetupScale( matrix_t m, vec_t x, vec_t y, vec_t z );
-	void     MatrixSetupShear( matrix_t m, vec_t x, vec_t y );
-	void     MatrixMultiply( const matrix_t a, const matrix_t b, matrix_t out );
-	void     MatrixMultiply2( matrix_t m, const matrix_t m2 );
-	void     MatrixMultiplyRotation( matrix_t m, vec_t pitch, vec_t yaw, vec_t roll );
-	void     MatrixMultiplyZRotation( matrix_t m, vec_t degrees );
-	void     MatrixMultiplyTranslation( matrix_t m, vec_t x, vec_t y, vec_t z );
-	void     MatrixMultiplyScale( matrix_t m, vec_t x, vec_t y, vec_t z );
-	void     MatrixMultiplyShear( matrix_t m, vec_t x, vec_t y );
-	void     MatrixToAngles( const matrix_t m, vec3_t angles );
-	void     MatrixFromAngles( matrix_t m, vec_t pitch, vec_t yaw, vec_t roll );
-	void     MatrixFromVectorsFLU( matrix_t m, const vec3_t forward, const vec3_t left, const vec3_t up );
-	void     MatrixFromVectorsFRU( matrix_t m, const vec3_t forward, const vec3_t right, const vec3_t up );
-	void     MatrixFromQuat( matrix_t m, const quat_t q );
-	void     MatrixFromPlanes( matrix_t m, const vec4_t left, const vec4_t right, const vec4_t bottom, const vec4_t top,
-	                           const vec4_t near, const vec4_t far );
-	void     MatrixToVectorsFLU( const matrix_t m, vec3_t forward, vec3_t left, vec3_t up );
-	void     MatrixToVectorsFRU( const matrix_t m, vec3_t forward, vec3_t right, vec3_t up );
-	void     MatrixSetupTransformFromVectorsFLU( matrix_t m, const vec3_t forward, const vec3_t left, const vec3_t up, const vec3_t origin );
-	void     MatrixSetupTransformFromVectorsFRU( matrix_t m, const vec3_t forward, const vec3_t right, const vec3_t up, const vec3_t origin );
-	void     MatrixSetupTransformFromRotation( matrix_t m, const matrix_t rot, const vec3_t origin );
-	void     MatrixSetupTransformFromQuat( matrix_t m, const quat_t quat, const vec3_t origin );
-	void     MatrixAffineInverse( const matrix_t in, matrix_t out );
-	void     MatrixTransformNormal( const matrix_t m, const vec3_t in, vec3_t out );
-	void     MatrixTransformNormal2( const matrix_t m, vec3_t inout );
-	void     MatrixTransformPoint( const matrix_t m, const vec3_t in, vec3_t out );
-	void     MatrixTransformPoint2( const matrix_t m, vec3_t inout );
-	void     MatrixTransform4( const matrix_t m, const vec4_t in, vec4_t out );
-	void     MatrixTransformPlane( const matrix_t m, const vec4_t in, vec4_t out );
-	void     MatrixTransformPlane2( const matrix_t m, vec4_t inout );
-	void     MatrixTransformBounds( const matrix_t m, const vec3_t mins, const vec3_t maxs, vec3_t omins, vec3_t omaxs );
-	void     MatrixPerspectiveProjection( matrix_t m, vec_t left, vec_t right, vec_t bottom, vec_t top, vec_t near, vec_t far );
-	void     MatrixPerspectiveProjectionLH( matrix_t m, vec_t left, vec_t right, vec_t bottom, vec_t top, vec_t near, vec_t far );
-	void     MatrixPerspectiveProjectionRH( matrix_t m, vec_t left, vec_t right, vec_t bottom, vec_t top, vec_t near, vec_t far );
-	void     MatrixPerspectiveProjectionFovYAspectLH( matrix_t m, vec_t fov, vec_t aspect, vec_t near, vec_t far );
-	void     MatrixPerspectiveProjectionFovXYLH( matrix_t m, vec_t fovX, vec_t fovY, vec_t near, vec_t far );
-	void     MatrixPerspectiveProjectionFovXYRH( matrix_t m, vec_t fovX, vec_t fovY, vec_t near, vec_t far );
-	void     MatrixPerspectiveProjectionFovXYInfiniteRH( matrix_t m, vec_t fovX, vec_t fovY, vec_t near );
-	void     MatrixOrthogonalProjection( matrix_t m, vec_t left, vec_t right, vec_t bottom, vec_t top, vec_t near, vec_t far );
-
-	void     MatrixOrthogonalProjectionLH( matrix_t m, vec_t left, vec_t right, vec_t bottom, vec_t top, vec_t near, vec_t far );
-	void     MatrixOrthogonalProjectionRH( matrix_t m, vec_t left, vec_t right, vec_t bottom, vec_t top, vec_t near, vec_t far );
-
-	void     MatrixPlaneReflection( matrix_t m, const vec4_t plane );
-
-	void     MatrixLookAtLH( matrix_t output, const vec3_t pos, const vec3_t dir, const vec3_t up );
-	void     MatrixLookAtRH( matrix_t m, const vec3_t eye, const vec3_t dir, const vec3_t up );
-	void     MatrixScaleTranslateToUnitCube( matrix_t m, const vec3_t mins, const vec3_t maxs );
-	void     MatrixCrop( matrix_t m, const vec3_t mins, const vec3_t maxs );
-
-	inline void AnglesToMatrix( const vec3_t angles, matrix_t m )
-	{
-		MatrixFromAngles( m, angles[ PITCH ], angles[ YAW ], angles[ ROLL ] );
-	}
-
-//=============================================
-
-// RB: XreaL quaternion math functions required by the renderer
-
-#define QuatSet(q,x,y,z,w) (( q )[ 0 ] = ( x ),( q )[ 1 ] = ( y ),( q )[ 2 ] = ( z ),( q )[ 3 ] = ( w ))
-#define QuatCopy(a,b)      (( b )[ 0 ] = ( a )[ 0 ],( b )[ 1 ] = ( a )[ 1 ],( b )[ 2 ] = ( a )[ 2 ],( b )[ 3 ] = ( a )[ 3 ] )
-
-#define QuatCompare(a,b)   (( a )[ 0 ] == ( b )[ 0 ] && ( a )[ 1 ] == ( b )[ 1 ] && ( a )[ 2 ] == ( b )[ 2 ] && ( a )[ 3 ] == ( b )[ 3 ] )
-
-	inline void QuatClear( quat_t q )
-	{
-		q[ 0 ] = 0;
-		q[ 1 ] = 0;
-		q[ 2 ] = 0;
-		q[ 3 ] = 1;
-	}
-
-	inline void QuatZero( quat_t o )
-	{
-		o[ 0 ] = 0.0f;
-		o[ 1 ] = 0.0f;
-		o[ 2 ] = 0.0f;
-		o[ 3 ] = 0.0f;
-	}
-
-	inline void QuatAdd( const quat_t p, const quat_t q,
-				    quat_t o )
-	{
-		o[ 0 ] = p[ 0 ] + q[ 0 ];
-		o[ 1 ] = p[ 1 ] + q[ 1 ];
-		o[ 2 ] = p[ 2 ] + q[ 2 ];
-		o[ 3 ] = p[ 3 ] + q[ 3 ];
-	}
-
-	inline void QuatMA( const quat_t p, float f, const quat_t q,
-				   quat_t o )
-	{
-		o[ 0 ] = p[ 0 ] + f * q[ 0 ];
-		o[ 1 ] = p[ 1 ] + f * q[ 1 ];
-		o[ 2 ] = p[ 2 ] + f * q[ 2 ];
-		o[ 3 ] = p[ 3 ] + f * q[ 3 ];
-	}
-
-	inline void QuatCalcW( quat_t q )
-	{
-		vec_t term = 1.0f - ( q[ 0 ] * q[ 0 ] + q[ 1 ] * q[ 1 ] + q[ 2 ] * q[ 2 ] );
-
-		if ( term < 0.0 )
-		{
-			q[ 3 ] = 0.0;
-		}
-		else
-		{
-			q[ 3 ] = -sqrt( term );
-		}
-	}
-
-	inline void QuatInverse( quat_t q )
-	{
-		q[ 0 ] = -q[ 0 ];
-		q[ 1 ] = -q[ 1 ];
-		q[ 2 ] = -q[ 2 ];
-	}
-
-	inline void QuatAntipodal( quat_t q )
-	{
-		q[ 0 ] = -q[ 0 ];
-		q[ 1 ] = -q[ 1 ];
-		q[ 2 ] = -q[ 2 ];
-		q[ 3 ] = -q[ 3 ];
-	}
-
-	inline vec_t QuatLength( const quat_t q )
-	{
-		return ( vec_t ) sqrt( q[ 0 ] * q[ 0 ] + q[ 1 ] * q[ 1 ] + q[ 2 ] * q[ 2 ] + q[ 3 ] * q[ 3 ] );
-	}
-
-	vec_t QuatNormalize( quat_t q );
-
-	void  QuatFromAngles( quat_t q, vec_t pitch, vec_t yaw, vec_t roll );
-
-	inline void AnglesToQuat( const vec3_t angles, quat_t q )
-	{
-		QuatFromAngles( q, angles[ PITCH ], angles[ YAW ], angles[ ROLL ] );
-	}
-
-	void QuatFromMatrix( quat_t q, const matrix_t m );
-	void QuatToVectorsFLU( const quat_t quat, vec3_t forward, vec3_t left, vec3_t up );
-	void QuatToVectorsFRU( const quat_t quat, vec3_t forward, vec3_t right, vec3_t up );
-
-	void QuatToAxis( const quat_t q, vec3_t axis[ 3 ] );
-	void QuatToAngles( const quat_t q, vec3_t angles );
-
-// Quaternion multiplication, analogous to the matrix multiplication routines.
-//
-// https://en.wikipedia.org/wiki/Quaternions_and_spatial_rotation
-// Two rotation quaternions can be combined into one equivalent quaternion by the relation:
-// q' = q2q1
-// in which q' corresponds to the rotation q1 followed by the rotation q2.
-
-	void QuatMultiply( const quat_t qa, const quat_t qb, quat_t qc );
-	void QuatMultiply2( quat_t qa, const quat_t qb);
-
-	void QuatSlerp( const quat_t from, const quat_t to, float frac, quat_t out );
-	void QuatTransformVector( const quat_t q, const vec3_t in, vec3_t out );
-	void QuatTransformVectorInverse( const quat_t q, const vec3_t in, vec3_t out );
-
-//=============================================
-// combining Transformations
-
-#if idx86_sse
-/* swizzles for _mm_shuffle_ps instruction */
-#define SWZ_XXXX 0x00
-#define SWZ_YXXX 0x01
-#define SWZ_ZXXX 0x02
-#define SWZ_WXXX 0x03
-#define SWZ_XYXX 0x04
-#define SWZ_YYXX 0x05
-#define SWZ_ZYXX 0x06
-#define SWZ_WYXX 0x07
-#define SWZ_XZXX 0x08
-#define SWZ_YZXX 0x09
-#define SWZ_ZZXX 0x0a
-#define SWZ_WZXX 0x0b
-#define SWZ_XWXX 0x0c
-#define SWZ_YWXX 0x0d
-#define SWZ_ZWXX 0x0e
-#define SWZ_WWXX 0x0f
-#define SWZ_XXYX 0x10
-#define SWZ_YXYX 0x11
-#define SWZ_ZXYX 0x12
-#define SWZ_WXYX 0x13
-#define SWZ_XYYX 0x14
-#define SWZ_YYYX 0x15
-#define SWZ_ZYYX 0x16
-#define SWZ_WYYX 0x17
-#define SWZ_XZYX 0x18
-#define SWZ_YZYX 0x19
-#define SWZ_ZZYX 0x1a
-#define SWZ_WZYX 0x1b
-#define SWZ_XWYX 0x1c
-#define SWZ_YWYX 0x1d
-#define SWZ_ZWYX 0x1e
-#define SWZ_WWYX 0x1f
-#define SWZ_XXZX 0x20
-#define SWZ_YXZX 0x21
-#define SWZ_ZXZX 0x22
-#define SWZ_WXZX 0x23
-#define SWZ_XYZX 0x24
-#define SWZ_YYZX 0x25
-#define SWZ_ZYZX 0x26
-#define SWZ_WYZX 0x27
-#define SWZ_XZZX 0x28
-#define SWZ_YZZX 0x29
-#define SWZ_ZZZX 0x2a
-#define SWZ_WZZX 0x2b
-#define SWZ_XWZX 0x2c
-#define SWZ_YWZX 0x2d
-#define SWZ_ZWZX 0x2e
-#define SWZ_WWZX 0x2f
-#define SWZ_XXWX 0x30
-#define SWZ_YXWX 0x31
-#define SWZ_ZXWX 0x32
-#define SWZ_WXWX 0x33
-#define SWZ_XYWX 0x34
-#define SWZ_YYWX 0x35
-#define SWZ_ZYWX 0x36
-#define SWZ_WYWX 0x37
-#define SWZ_XZWX 0x38
-#define SWZ_YZWX 0x39
-#define SWZ_ZZWX 0x3a
-#define SWZ_WZWX 0x3b
-#define SWZ_XWWX 0x3c
-#define SWZ_YWWX 0x3d
-#define SWZ_ZWWX 0x3e
-#define SWZ_WWWX 0x3f
-#define SWZ_XXXY 0x40
-#define SWZ_YXXY 0x41
-#define SWZ_ZXXY 0x42
-#define SWZ_WXXY 0x43
-#define SWZ_XYXY 0x44
-#define SWZ_YYXY 0x45
-#define SWZ_ZYXY 0x46
-#define SWZ_WYXY 0x47
-#define SWZ_XZXY 0x48
-#define SWZ_YZXY 0x49
-#define SWZ_ZZXY 0x4a
-#define SWZ_WZXY 0x4b
-#define SWZ_XWXY 0x4c
-#define SWZ_YWXY 0x4d
-#define SWZ_ZWXY 0x4e
-#define SWZ_WWXY 0x4f
-#define SWZ_XXYY 0x50
-#define SWZ_YXYY 0x51
-#define SWZ_ZXYY 0x52
-#define SWZ_WXYY 0x53
-#define SWZ_XYYY 0x54
-#define SWZ_YYYY 0x55
-#define SWZ_ZYYY 0x56
-#define SWZ_WYYY 0x57
-#define SWZ_XZYY 0x58
-#define SWZ_YZYY 0x59
-#define SWZ_ZZYY 0x5a
-#define SWZ_WZYY 0x5b
-#define SWZ_XWYY 0x5c
-#define SWZ_YWYY 0x5d
-#define SWZ_ZWYY 0x5e
-#define SWZ_WWYY 0x5f
-#define SWZ_XXZY 0x60
-#define SWZ_YXZY 0x61
-#define SWZ_ZXZY 0x62
-#define SWZ_WXZY 0x63
-#define SWZ_XYZY 0x64
-#define SWZ_YYZY 0x65
-#define SWZ_ZYZY 0x66
-#define SWZ_WYZY 0x67
-#define SWZ_XZZY 0x68
-#define SWZ_YZZY 0x69
-#define SWZ_ZZZY 0x6a
-#define SWZ_WZZY 0x6b
-#define SWZ_XWZY 0x6c
-#define SWZ_YWZY 0x6d
-#define SWZ_ZWZY 0x6e
-#define SWZ_WWZY 0x6f
-#define SWZ_XXWY 0x70
-#define SWZ_YXWY 0x71
-#define SWZ_ZXWY 0x72
-#define SWZ_WXWY 0x73
-#define SWZ_XYWY 0x74
-#define SWZ_YYWY 0x75
-#define SWZ_ZYWY 0x76
-#define SWZ_WYWY 0x77
-#define SWZ_XZWY 0x78
-#define SWZ_YZWY 0x79
-#define SWZ_ZZWY 0x7a
-#define SWZ_WZWY 0x7b
-#define SWZ_XWWY 0x7c
-#define SWZ_YWWY 0x7d
-#define SWZ_ZWWY 0x7e
-#define SWZ_WWWY 0x7f
-#define SWZ_XXXZ 0x80
-#define SWZ_YXXZ 0x81
-#define SWZ_ZXXZ 0x82
-#define SWZ_WXXZ 0x83
-#define SWZ_XYXZ 0x84
-#define SWZ_YYXZ 0x85
-#define SWZ_ZYXZ 0x86
-#define SWZ_WYXZ 0x87
-#define SWZ_XZXZ 0x88
-#define SWZ_YZXZ 0x89
-#define SWZ_ZZXZ 0x8a
-#define SWZ_WZXZ 0x8b
-#define SWZ_XWXZ 0x8c
-#define SWZ_YWXZ 0x8d
-#define SWZ_ZWXZ 0x8e
-#define SWZ_WWXZ 0x8f
-#define SWZ_XXYZ 0x90
-#define SWZ_YXYZ 0x91
-#define SWZ_ZXYZ 0x92
-#define SWZ_WXYZ 0x93
-#define SWZ_XYYZ 0x94
-#define SWZ_YYYZ 0x95
-#define SWZ_ZYYZ 0x96
-#define SWZ_WYYZ 0x97
-#define SWZ_XZYZ 0x98
-#define SWZ_YZYZ 0x99
-#define SWZ_ZZYZ 0x9a
-#define SWZ_WZYZ 0x9b
-#define SWZ_XWYZ 0x9c
-#define SWZ_YWYZ 0x9d
-#define SWZ_ZWYZ 0x9e
-#define SWZ_WWYZ 0x9f
-#define SWZ_XXZZ 0xa0
-#define SWZ_YXZZ 0xa1
-#define SWZ_ZXZZ 0xa2
-#define SWZ_WXZZ 0xa3
-#define SWZ_XYZZ 0xa4
-#define SWZ_YYZZ 0xa5
-#define SWZ_ZYZZ 0xa6
-#define SWZ_WYZZ 0xa7
-#define SWZ_XZZZ 0xa8
-#define SWZ_YZZZ 0xa9
-#define SWZ_ZZZZ 0xaa
-#define SWZ_WZZZ 0xab
-#define SWZ_XWZZ 0xac
-#define SWZ_YWZZ 0xad
-#define SWZ_ZWZZ 0xae
-#define SWZ_WWZZ 0xaf
-#define SWZ_XXWZ 0xb0
-#define SWZ_YXWZ 0xb1
-#define SWZ_ZXWZ 0xb2
-#define SWZ_WXWZ 0xb3
-#define SWZ_XYWZ 0xb4
-#define SWZ_YYWZ 0xb5
-#define SWZ_ZYWZ 0xb6
-#define SWZ_WYWZ 0xb7
-#define SWZ_XZWZ 0xb8
-#define SWZ_YZWZ 0xb9
-#define SWZ_ZZWZ 0xba
-#define SWZ_WZWZ 0xbb
-#define SWZ_XWWZ 0xbc
-#define SWZ_YWWZ 0xbd
-#define SWZ_ZWWZ 0xbe
-#define SWZ_WWWZ 0xbf
-#define SWZ_XXXW 0xc0
-#define SWZ_YXXW 0xc1
-#define SWZ_ZXXW 0xc2
-#define SWZ_WXXW 0xc3
-#define SWZ_XYXW 0xc4
-#define SWZ_YYXW 0xc5
-#define SWZ_ZYXW 0xc6
-#define SWZ_WYXW 0xc7
-#define SWZ_XZXW 0xc8
-#define SWZ_YZXW 0xc9
-#define SWZ_ZZXW 0xca
-#define SWZ_WZXW 0xcb
-#define SWZ_XWXW 0xcc
-#define SWZ_YWXW 0xcd
-#define SWZ_ZWXW 0xce
-#define SWZ_WWXW 0xcf
-#define SWZ_XXYW 0xd0
-#define SWZ_YXYW 0xd1
-#define SWZ_ZXYW 0xd2
-#define SWZ_WXYW 0xd3
-#define SWZ_XYYW 0xd4
-#define SWZ_YYYW 0xd5
-#define SWZ_ZYYW 0xd6
-#define SWZ_WYYW 0xd7
-#define SWZ_XZYW 0xd8
-#define SWZ_YZYW 0xd9
-#define SWZ_ZZYW 0xda
-#define SWZ_WZYW 0xdb
-#define SWZ_XWYW 0xdc
-#define SWZ_YWYW 0xdd
-#define SWZ_ZWYW 0xde
-#define SWZ_WWYW 0xdf
-#define SWZ_XXZW 0xe0
-#define SWZ_YXZW 0xe1
-#define SWZ_ZXZW 0xe2
-#define SWZ_WXZW 0xe3
-#define SWZ_XYZW 0xe4
-#define SWZ_YYZW 0xe5
-#define SWZ_ZYZW 0xe6
-#define SWZ_WYZW 0xe7
-#define SWZ_XZZW 0xe8
-#define SWZ_YZZW 0xe9
-#define SWZ_ZZZW 0xea
-#define SWZ_WZZW 0xeb
-#define SWZ_XWZW 0xec
-#define SWZ_YWZW 0xed
-#define SWZ_ZWZW 0xee
-#define SWZ_WWZW 0xef
-#define SWZ_XXWW 0xf0
-#define SWZ_YXWW 0xf1
-#define SWZ_ZXWW 0xf2
-#define SWZ_WXWW 0xf3
-#define SWZ_XYWW 0xf4
-#define SWZ_YYWW 0xf5
-#define SWZ_ZYWW 0xf6
-#define SWZ_WYWW 0xf7
-#define SWZ_XZWW 0xf8
-#define SWZ_YZWW 0xf9
-#define SWZ_ZZWW 0xfa
-#define SWZ_WZWW 0xfb
-#define SWZ_XWWW 0xfc
-#define SWZ_YWWW 0xfd
-#define SWZ_ZWWW 0xfe
-#define SWZ_WWWW 0xff
-#define sseSwizzle( a, mask ) _mm_shuffle_ps( (a), (a), SWZ_##mask )
-
-	inline __m128 unitQuat() {
-		return _mm_set_ps( 1.0f, 0.0f, 0.0f, 0.0f ); // order is reversed
-	}
-	inline __m128 sseLoadInts( const int vec[4] ) {
-		return *(__m128 *)vec;
-	}
-	inline __m128 mask_0000() {
-		static const ALIGNED(16, int vec[4]) = {  0,  0,  0,  0 };
-		return sseLoadInts( vec );
-	}
-	inline __m128 mask_000W() {
-		static const ALIGNED(16, int vec[4]) = {  0,  0,  0, -1 };
-		return sseLoadInts( vec );
-	}
-	inline __m128 mask_XYZ0() {
-		static const ALIGNED(16, int vec[4]) = { -1, -1, -1,  0 };
-		return sseLoadInts( vec );
-	}
-
-	inline __m128 sign_000W() {
-		static const ALIGNED(16, int vec[4]) = { 0, 0, 0, 1<<31 };
-		return sseLoadInts( vec );
-	}
-	inline __m128 sign_XYZ0() {
-		static const ALIGNED(16, int vec[4]) = { 1<<31, 1<<31, 1<<31,  0 };
-		return sseLoadInts( vec );
-	}
-	inline __m128 sign_XYZW() {
-		static const ALIGNED(16, int vec[4]) = { 1<<31, 1<<31, 1<<31, 1<<31 };
-		return sseLoadInts( vec );
-	}
-
-	inline __m128 sseDot4( __m128 a, __m128 b ) {
-		__m128 prod = _mm_mul_ps( a, b );
-		__m128 sum1 = _mm_add_ps( prod, sseSwizzle( prod, YXWZ ) );
-		__m128 sum2 = _mm_add_ps( sum1, sseSwizzle( sum1, ZWXY ) );
-		return sum2;
-	}
-	inline __m128 sseCrossProduct( __m128 a, __m128 b ) {
-		__m128 a_yzx = sseSwizzle( a, YZXW );
-		__m128 b_yzx = sseSwizzle( b, YZXW );
-		__m128 c_zxy = _mm_sub_ps( _mm_mul_ps( a, b_yzx ),
-					   _mm_mul_ps( a_yzx, b ) );
-		return sseSwizzle( c_zxy, YZXW );
-	}
-	inline __m128 sseQuatMul( __m128 a, __m128 b ) {
-		__m128 a1 = sseSwizzle( a, WWWW );
-		__m128 c1 = _mm_mul_ps( a1, b );
-		__m128 a2 = sseSwizzle( a, XYZX );
-		__m128 b2 = sseSwizzle( b, WWWX );
-		__m128 c2 = _mm_xor_ps( _mm_mul_ps(a2, b2), sign_000W() );
-		__m128 a3 = sseSwizzle( a, YZXY );
-		__m128 b3 = sseSwizzle( b, ZXYY );
-		__m128 c3 = _mm_xor_ps( _mm_mul_ps(a3, b3), sign_000W() );
-		__m128 a4 = sseSwizzle( a, ZXYZ);
-		__m128 b4 = sseSwizzle( b, YZXZ);
-		__m128 c4 = _mm_mul_ps( a4, b4 );
-		return _mm_add_ps( _mm_add_ps(c1, c2), _mm_sub_ps(c3, c4) );
-	}
-	inline __m128 sseQuatNormalize( __m128 q ) {
-		__m128 p = _mm_mul_ps( q, q );
-		__m128 t, h;
-		p = _mm_add_ps( sseSwizzle( p, XXZZ ),
-				sseSwizzle( p, YYWW ) );
-		p = _mm_add_ps( sseSwizzle( p, XXXX ),
-				sseSwizzle( p, ZZZZ ) );
-		t = _mm_rsqrt_ps( p );
-		h = _mm_mul_ps( _mm_set1_ps( 0.5f ), t );
-		t = _mm_mul_ps( _mm_mul_ps( t, t ), p );
-		t = _mm_sub_ps( _mm_set1_ps( 3.0f ), t );
-		t = _mm_mul_ps( h, t );
-		return _mm_mul_ps( q, t );
-	}
-	inline __m128 sseQuatTransform( __m128 q, __m128 vec ) {
-		__m128 t, t2;
-		t = sseCrossProduct( q, vec );
-		t = _mm_add_ps( t, t );
-		t2 = sseCrossProduct( q, t );
-		t = _mm_mul_ps( sseSwizzle( q, WWWW ), t );
-		return _mm_add_ps( _mm_add_ps( vec, t2 ), t );
-	}
-	inline __m128 sseQuatTransformInverse( __m128 q, __m128 vec ) {
-		__m128 t, t2;
-		t = sseCrossProduct( vec, q );
-		t = _mm_add_ps( t, t );
-		t2 = sseCrossProduct( t, q );
-		t = _mm_mul_ps( sseSwizzle( q, WWWW ), t );
-		return _mm_add_ps( _mm_add_ps( vec, t2 ), t );
-	}
-	inline __m128 sseLoadVec3( const vec3_t vec ) {
-		__m128 v = _mm_load_ss( &vec[ 2 ] );
-		v = sseSwizzle( v, YYXY );
-		v = _mm_loadl_pi( v, (__m64 *)vec );
-		return v;
-	}
-	inline void sseStoreVec3( __m128 in, vec3_t out ) {
-		_mm_storel_pi( (__m64 *)out, in );
-		__m128 v = sseSwizzle( in, ZZZZ );
-		_mm_store_ss( &out[ 2 ], v );
-	}
-	inline void TransInit( transform_t *t ) {
-		__m128 u = unitQuat();
-		t->sseRot = u;
-		t->sseTransScale = u;
-	}
-	inline void TransCopy( const transform_t *in, transform_t *out ) {
-		out->sseRot = in->sseRot;
-		out->sseTransScale = in->sseTransScale;
-	}
-	inline void TransformPoint( const transform_t *t,
-					   const vec3_t in, vec3_t out ) {
-		__m128 ts = t->sseTransScale;
-		__m128 tmp = sseQuatTransform( t->sseRot, _mm_loadu_ps( in ) );
-		tmp = _mm_mul_ps( tmp, sseSwizzle( ts, WWWW ) );
-		tmp = _mm_add_ps( tmp, ts );
-		sseStoreVec3( tmp, out );
-	}
-	inline void TransformPointInverse( const transform_t *t,
-						  const vec3_t in, vec3_t out ) {
-		__m128 ts = t->sseTransScale;
-		__m128 v = _mm_sub_ps( _mm_loadu_ps( in ), ts );
-		v = _mm_mul_ps( v, _mm_rcp_ps( sseSwizzle( ts, WWWW ) ) );
-		v = sseQuatTransformInverse( t->sseRot, v );
-		sseStoreVec3( v, out );
-	}
-	inline void TransformNormalVector( const transform_t *t,
-						  const vec3_t in, vec3_t out ) {
-		__m128 v = _mm_loadu_ps( in );
-		v = sseQuatTransform( t->sseRot, v );
-		sseStoreVec3( v, out );
-	}
-	inline void TransformNormalVectorInverse( const transform_t *t,
-							 const vec3_t in, vec3_t out ) {
-		__m128 v = _mm_loadu_ps( in );
-		v = sseQuatTransformInverse( t->sseRot, v );
-		sseStoreVec3( v, out );
-	}
-	inline __m128 sseAxisAngleToQuat( const vec3_t axis, float angle ) {
-		__m128 sa = _mm_set1_ps( sin( 0.5f * angle ) );
-		__m128 ca = _mm_set1_ps( cos( 0.5f * angle ) );
-		__m128 a = _mm_loadu_ps( axis );
-		a = _mm_and_ps( a, mask_XYZ0() );
-		a = _mm_mul_ps( a, sa );
-		return _mm_or_ps( a, _mm_and_ps( ca, mask_000W() ) );
-	}
-	inline void TransInitRotationQuat( const quat_t quat,
-						  transform_t *t ) {
-		t->sseRot = _mm_loadu_ps( quat );
-		t->sseTransScale = unitQuat();
-	}
-	inline void TransInitRotation( const vec3_t axis, float angle,
-					      transform_t *t ) {
-		t->sseRot = sseAxisAngleToQuat( axis, angle );
-		t->sseTransScale = unitQuat();
-	}
-	inline void TransInitTranslation( const vec3_t vec, transform_t *t ) {
-		__m128 v = _mm_loadu_ps( vec );
-		v = _mm_and_ps( v, mask_XYZ0() );
-		t->sseRot = unitQuat();
-		t->sseTransScale = _mm_or_ps( v, unitQuat() );
-	}
-	inline void TransInitScale( float factor, transform_t *t ) {
-		__m128 f = _mm_set1_ps( factor );
-		f = _mm_and_ps( f, mask_000W() );
-		t->sseRot = unitQuat();
-		t->sseTransScale = f;
-	}
-	inline void TransInsRotationQuat( const quat_t quat, transform_t *t ) {
-		__m128 q = _mm_loadu_ps( quat );
-		t->sseRot = sseQuatMul( t->sseRot, q );
-	}
-	inline void TransInsRotation( const vec3_t axis, float angle,
-					     transform_t *t ) {
-		__m128 q = sseAxisAngleToQuat( axis, angle );
-		t->sseRot = sseQuatMul( q, t->sseRot );
-	}
-	inline void TransAddRotationQuat( const quat_t quat, transform_t *t ) {
-		__m128 q = _mm_loadu_ps( quat );
-		__m128 transformed = sseQuatTransform( q, t->sseTransScale );
-		t->sseRot = sseQuatMul( q, t->sseRot );
-		t->sseTransScale = _mm_or_ps( _mm_and_ps( transformed, mask_XYZ0() ),
-					      _mm_and_ps( t->sseTransScale, mask_000W() ) );
-	}
-	inline void TransAddRotation( const vec3_t axis, float angle,
-					     transform_t *t ) {
-		__m128 q = sseAxisAngleToQuat( axis, angle );
-		__m128 transformed = sseQuatTransform( q, t->sseTransScale );
-		t->sseRot = sseQuatMul( t->sseRot, q );
-		t->sseTransScale = _mm_or_ps( _mm_and_ps( transformed, mask_XYZ0() ),
-					      _mm_and_ps( t->sseTransScale, mask_000W() ) );
-	}
-	inline void TransInsScale( float factor, transform_t *t ) {
-		t->scale *= factor;
-	}
-	inline void TransAddScale( float factor, transform_t *t ) {
-		__m128 f = _mm_set1_ps( factor );
-		t->sseTransScale = _mm_mul_ps( f, t->sseTransScale );
-	}
-	inline void TransInsTranslation( const vec3_t vec,
-						transform_t *t ) {
-		__m128 v = _mm_loadu_ps( vec );
-		__m128 ts = t->sseTransScale;
-		v = sseQuatTransform( t->sseRot, v );
-		v = _mm_mul_ps( v, sseSwizzle( ts, WWWW ) );
-		v = _mm_and_ps( v, mask_XYZ0() );
-		t->sseTransScale = _mm_add_ps( ts, v );
-	}
-	inline void TransAddTranslation( const vec3_t vec,
-						transform_t *t ) {
-		__m128 v = _mm_loadu_ps( vec );
-		v = _mm_and_ps( v, mask_XYZ0() );
-		t->sseTransScale = _mm_add_ps( t->sseTransScale, v );
-	}
-	inline void TransCombine( const transform_t *a,
-					 const transform_t *b,
-					 transform_t *out ) {
-		__m128 aRot = a->sseRot;
-		__m128 aTS = a->sseTransScale;
-		__m128 bRot = b->sseRot;
-		__m128 bTS = b->sseTransScale;
-		__m128 tmp = sseQuatTransform( bRot, aTS );
-		tmp = _mm_or_ps( _mm_and_ps( tmp, mask_XYZ0() ),
-				 _mm_and_ps( aTS, mask_000W() ) );
-		tmp = _mm_mul_ps( tmp, sseSwizzle( bTS, WWWW ) );
-		out->sseTransScale = _mm_add_ps( tmp, _mm_and_ps( bTS, mask_XYZ0() ) );
-		out->sseRot = sseQuatMul( bRot, aRot );
-	}
-	inline void TransInverse( const transform_t *in,
-					 transform_t *out ) {
-		__m128 rot = in->sseRot;
-		__m128 ts = in->sseTransScale;
-		__m128 invS = _mm_rcp_ps( sseSwizzle( ts, WWWW ) );
-		__m128 invRot = _mm_xor_ps( rot, sign_XYZ0() );
-		__m128 invT = _mm_xor_ps( ts, sign_XYZ0() );
-		__m128 tmp = sseQuatTransform( invRot, invT );
-		tmp = _mm_mul_ps( tmp, invS );
-		out->sseRot = invRot;
-		out->sseTransScale = _mm_or_ps( _mm_and_ps( tmp, mask_XYZ0() ),
-						_mm_and_ps( invS, mask_000W() ) );
-	}
-	inline void TransStartLerp( transform_t *t ) {
-		t->sseRot = mask_0000();
-		t->sseTransScale = mask_0000();
-	}
-	inline void TransAddWeight( float weight, const transform_t *a,
-					   transform_t *out ) {
-		__m128 w = _mm_set1_ps( weight );
-		__m128 d = sseDot4( a->sseRot, out->sseRot );
-		out->sseTransScale = _mm_add_ps( out->sseTransScale,
-						 _mm_mul_ps( w, a->sseTransScale ) );
-		w = _mm_xor_ps( w, _mm_and_ps( d, sign_XYZW() ) );
-		out->sseRot = _mm_add_ps( out->sseRot,
-					  _mm_mul_ps( w, a->sseRot ) );
-}
-	inline void TransEndLerp( transform_t *t ) {
-		t->sseRot = sseQuatNormalize( t->sseRot );
-	}
-#else
-	void TransInit( transform_t *t );
-	void TransCopy( const transform_t *in, transform_t *out );
-
-	void TransformPoint( const transform_t *t, const vec3_t in, vec3_t out );
-	void TransformPointInverse( const transform_t *t, const vec3_t in, vec3_t out );
-	void TransformNormalVector( const transform_t *t, const vec3_t in, vec3_t out );
-	void TransformNormalVectorInverse( const transform_t *t, const vec3_t in, vec3_t out );
-
-	void TransInitRotationQuat( const quat_t quat, transform_t *t );
-	void TransInitRotation( const vec3_t axis, float angle,
-				transform_t *t );
-	void TransInitTranslation( const vec3_t vec, transform_t *t );
-	void TransInitScale( float factor, transform_t *t );
-
-	void TransInsRotationQuat( const quat_t quat, transform_t *t );
-	void TransInsRotation( const vec3_t axis, float angle, transform_t *t );
-	void TransAddRotationQuat( const quat_t quat, transform_t *t );
-	void TransAddRotation( const vec3_t axis, float angle, transform_t *t );
-	void TransInsScale( float factor, transform_t *t );
-	void TransAddScale( float factor, transform_t *t );
-	void TransInsTranslation( const vec3_t vec, transform_t *t );
-	void TransAddTranslation( const vec3_t vec, transform_t *t );
-
-	void TransCombine( const transform_t *a, const transform_t *b,
-			   transform_t *c );
-	void TransInverse( const transform_t *in, transform_t *out );
-
-	void TransStartLerp( transform_t *t );
-	void TransAddWeight( float weight, const transform_t *a, transform_t *t );
-	void TransEndLerp( transform_t *t );
-#endif
+#include "q_math.h"
 
 //=============================================
 
@@ -1557,82 +443,7 @@ void         ByteToDir( int b, vec3_t dir );
 		char         string[ MAX_CVAR_VALUE_STRING ];
 	};
 
-	/*
-	==============================================================
-
-	COLLISION DETECTION
-
-	==============================================================
-	*/
-
 #include "SurfaceFlags.h" // shared with the q3map utility
-
-// plane types are used to speed some tests
-// 0-2 are axial planes
-#define PLANE_X          0
-#define PLANE_Y          1
-#define PLANE_Z          2
-#define PLANE_NON_AXIAL  3
-#define PLANE_NON_PLANAR 4
-
-	/*
-	=================
-	PlaneTypeForNormal
-	=================
-	*/
-
-#define PlaneTypeForNormal( x ) ( x[ 0 ] == 1.0 ? PLANE_X : ( x[ 1 ] == 1.0 ? PLANE_Y : ( x[ 2 ] == 1.0 ? PLANE_Z : ( x[ 0 ] == 0.f && x[ 1 ] == 0.f && x[ 2 ] == 0.f ? PLANE_NON_PLANAR : PLANE_NON_AXIAL ) ) ) )
-
-// plane_t structure
-	struct cplane_t
-	{
-		vec3_t normal;
-		float  dist;
-		byte   type; // for fast side tests: 0,1,2 = axial, 3 = nonaxial
-		byte   signbits; // signx + (signy<<1) + (signz<<2), used as lookup during collision
-		byte   pad[ 2 ];
-	};
-
-	enum class traceType_t
-	{
-	  TT_NONE,
-
-	  TT_AABB,
-	  TT_CAPSULE,
-	  TT_BISPHERE,
-
-	  TT_NUM_TRACE_TYPES
-	};
-
-// a trace is returned when a box is swept through the world
-	struct trace_t
-	{
-		bool allsolid; // if true, plane is not valid
-		bool startsolid; // if true, the initial point was in a solid area
-		float    fraction; // time completed, 1.0 = didn't hit anything
-		vec3_t   endpos; // final position
-		cplane_t plane; // surface normal at impact, transformed to world space
-		int      surfaceFlags; // surface hit
-		int      contents; // contents on other side of surface hit
-		int      entityNum; // entity the contacted surface is a part of
-		float    lateralFraction; // fraction of collision tangetially to the trace direction
-	};
-
-// trace->entityNum can also be 0 to (MAX_GENTITIES-1)
-// or ENTITYNUM_NONE, ENTITYNUM_WORLD
-
-// markfragments are returned by CM_MarkFragments()
-	struct markFragment_t
-	{
-		int firstPoint;
-		int numPoints;
-	};
-
-	struct orientation_t
-	{
-		vec3_t origin;
-		vec3_t axis[ 3 ];
-	};
 
 //=====================================================================
 
@@ -1669,9 +480,6 @@ void         ByteToDir( int b, vec3_t dir );
 	========================================================================
 	*/
 #define ANIM_BITS 10
-
-#define ANGLE2SHORT( x ) ( (int)( ( x ) * 65536 / 360 ) & 65535 )
-#define SHORT2ANGLE( x ) ( ( x ) * ( 360.0 / 65536 ) )
 
 #define SNAPFLAG_RATE_DELAYED 1
 #define SNAPFLAG_NOT_ACTIVE   2 // snapshot used during connection and for zombies
@@ -2054,8 +862,6 @@ struct fontInfo_t
 	float         glyphScale;
 	char          name[ MAX_QPATH ];
 };
-
-#define Square( x ) ( ( x ) * ( x ) )
 
 // real time
 //=============================================
